requirements.txt
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.27.0
autoawq>=0.2.5
numpy>=1.24.0
scipy>=1.11.0
tqdm>=4.66.0
sentencepiece>=0.1.99
bitsandbytes>=0.43.0
einops>=0.7.0
protobuf
tiktoken
